+++ library/src/main/java/candybar/lib/helpers/LocaleHelper.java
@@ -1,7 +1,8 @@
 package candybar.lib.helpers;
 
+import android.content.ComponentName;
 import android.content.Context;
-import android.content.pm.ApplicationInfo;
+import android.content.pm.ActivityInfo;
 import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.content.res.Resources;
@@ -99,23 +100,35 @@ public class LocaleHelper {
     }
 
     @Nullable
-    public static String getOtherAppLocaleName(@NonNull Context context, @NonNull Locale locale, @NonNull String packageName) {
+    public static String getOtherAppLocaleName(@NonNull Context context, @NonNull Locale locale, @NonNull String fullComponentName) {
         try {
-            PackageManager packageManager = context.getPackageManager();
-            ApplicationInfo info = packageManager.getApplicationInfo(packageName, PackageManager.GET_META_DATA);
+            int slashIndex = fullComponentName.indexOf("/");
+            String activityName = fullComponentName.substring(slashIndex).replace("/", "");
+            String packageName = fullComponentName.replace("/" + activityName, "");
+            ComponentName componentName = new ComponentName(packageName, activityName);
 
-            Resources res = packageManager.getResourcesForApplication(packageName);
+
+            Log.d("CandyBar", "Full Component Name: " + fullComponentName);
+            Log.d("CandyBar", "Activity Name: " + activityName);
+            Log.d("CandyBar", "Package Name: " + packageName);
+
+            PackageManager packageManager = context.getPackageManager();
+            ActivityInfo info = packageManager.getActivityInfo(componentName, PackageManager.GET_META_DATA);
+
+            Resources res = packageManager.getResourcesForActivity(componentName);
             Context otherAppContext = context.createPackageContext(packageName, Context.CONTEXT_IGNORE_SECURITY);
             Configuration configuration = new Configuration();
+
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
                 configuration = res.getConfiguration();
                 configuration.setLocale(locale);
                 return otherAppContext.createConfigurationContext(configuration).getString(info.labelRes);
             }
 
+
             configuration.locale = locale;
             res.updateConfiguration(configuration, context.getResources().getDisplayMetrics());
-            return res.getString(info.labelRes);
+            return info.loadLabel(packageManager).toString();
         } catch (Exception e) {
             LogUtil.e(Log.getStackTraceString(e));
         }