+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Option to Enable/Disable AppUpdater and Checking JSON Before Requesting.
 - Uses New Muzei API.
 - Now you can set Your Own Email Subject for Icon Requests.
+- Now it Loads Right Icon and Name from ComponentName instead of PackageName.
 
 # v3.5.0-b5
 - Updated Gradle Distribution and Gradle Plugin